+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [Unreleased]
 - Add macros for `console.log` functionalities for Huff contracts.
-  - This is experimental and will be maybe subject to breaking changes.
+  - See `test/console.t.huff` for examples.
 
 ## [1.0.0] - 2025-01-12
 - __Breaking__: Rename `HuffDeployer` to `HuffDeployerNeo` and `HuffConfig` to `HuffNeoConfig`